PyCharm 2025.2.4/requirements.txt
# 后端依赖包
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.8.1.78
h5py==3.10.0
astropy==6.0.0
astroml==0.4.2
openai==1.3.0
werkzeug==3.0.1
bcrypt==4.1.1
python-multipart==0.0.6

